Added abstract class 'LineTagHelper' for header and list item tags
